+++ scripts/bladebit_setup.sh
@@ -13,8 +13,9 @@ if [[ (${mode} =~ ^fullnode.*  || ${mode} =~ "plotter") && (${blockchains} == 'c
         cd / && echo "Installing bladebit on ${arch_name}..."
 
         # Now install CNI's separate binary for actual GPU plotting...
+        mkdir -p /opt/chia/bladebit
         cd /opt/chia/bladebit
-        rm -f ./bladebit  # Remove stale version bundled in the DEB package.
+        rm -f ./bladebit*  # Remove any stale versions bundled in the DEB package.
         if [[ "${arch_name}" == "x86_64" ]]; then
             curl -sLJO https://github.com/Chia-Network/bladebit/releases/download/v3.1.0/bladebit-v3.1.0-ubuntu-x86-64.tar.gz
             curl -sLJO https://github.com/Chia-Network/bladebit/releases/download/v3.1.0/bladebit-cuda-v3.1.0-ubuntu-x86-64.tar.gz