+++ scripts/forks/chia_launch.sh
@@ -35,7 +35,7 @@ if [[ "${blockchain_db_download}" == 'true' ]] \
   echo "Downloading Chia blockchain DB (many GBs in size) on first launch..."
   echo "Please be patient as this takes hours now, but saves days of syncing time later."
   mkdir -p /root/.chia/mainnet/db/chia && cd /root/.chia/mainnet/db/chia
-  # Latest Blockchain DB, first try direct download, then fallback to slower torrent
+  # Latest Blockchain DB via torrent download
   torrent=$(curl -s https://www.chia.net/downloads/ | grep -Po "https://torrents.chia.net/databases/mainnet/mainnet.\d{4}-\d{2}-\d{2}.tar.gz.torrent")
   echo "Please be patient! Downloading blockchain database indirectly (via libtorrent) from: "
   echo "    ${torrent}"
@@ -43,8 +43,8 @@ if [[ "${blockchain_db_download}" == 'true' ]] \
   /usr/bin/python /machinaris/scripts/chiadb_download.py $PWD/*.torrent >> /tmp/chiadb_download.log 2>&1
   echo "Now decompressing the blockchain database..."
   cd /root/.chia/mainnet/db/chia && tar -xf *.gz
+  mv blockchain_v2_mainnet.sqlite height-to-hash ..
   cd /root/.chia/mainnet/db
-  mv /root/.chia/mainnet/db/chia/blockchain_v2_mainnet.*.sqlite blockchain_v2_mainnet.sqlite
   rm -rf /root/.chia/mainnet/db/chia
 fi
 
@@ -99,6 +99,11 @@ if [[ ${mode} =~ ^fullnode.* ]]; then
     echo "Starting Chia Exporter service for Prometheus reporting..."
     sleep 20 && /usr/local/bin/chia-exporter serve 2>&1 > /root/.chia/mainnet/log/chia-exporter.log &
   fi
+  if [[ ${chia_data} == "true" ]]; then
+    echo "Starting Chia Data Layer services..."
+    sleep 20 && chia start data 2>&1 > /root/.chia/mainnet/log/chia-data.log &
+    sleep 20 && chia start data_layer_http 2>&1 >> /root/.chia/mainnet/log/chia-data.log &
+  fi
   if [[ ${mode} =~ .*timelord$ ]]; then
     if [ ! -f vdf_bench ]; then
         echo "Building timelord binaries..."
@@ -136,6 +141,10 @@ elif [[ ${mode} =~ ^harvester.* ]]; then
     chia configure --set-farmer-peer ${farmer_address}:${farmer_port}  2>&1 >> /root/.chia/mainnet/log/init.log
     chia configure --enable-upnp false  2>&1 >> /root/.chia/mainnet/log/init.log
     chia start harvester -r
+    if [[ ${chia_exporter} == "true" ]]; then
+      echo "Starting Chia Exporter service for Prometheus reporting..."
+      sleep 20 && /usr/local/bin/chia-exporter serve 2>&1 > /root/.chia/mainnet/log/chia-exporter.log &
+    fi
   fi
 elif [[ ${mode} == 'plotter' ]]; then
     echo "Starting in Plotter-only mode.  Run Plotman from either CLI or WebUI."