+++ scripts/forks/greenbtc_install.sh
@@ -4,8 +4,8 @@
 #
 
 GREENBTC_BRANCH=$1
-# On 2023-01-02
-HASH=e14bc12696af1cbcdf89a3e7bea0cdb1dd0e89ad
+# On 2024-02-16
+HASH=200816a51c24375fadcbcd25a0b1cb7549f61c77
 
 if [ -z ${GREENBTC_BRANCH} ]; then
     echo 'Skipping GreenBTC install as not requested.'